FDG-9959 Update download wrapper component name
